+++ contracts/axelar-operators/src/storage_types.rs
@@ -3,7 +3,6 @@ use soroban_sdk::{contracttype, Address};
 #[contracttype]
 #[derive(Clone, Debug)]
 pub enum DataKey {
-    Initialized,
     Owner,
     Operators(Address),
 }
